requirements.txt
Django==1.9
Pillow
MySQL-python
telepot
python3-netsnmp
